Refactor user permissions checks
- Remove duplicate function calls
- Refactor permission checks into new file
